+++ Runtime/AwaitExtensions.cs
@@ -11,6 +11,7 @@ namespace GameLovers.AsyncAwait
         public static TaskAwaiter<int> GetAwaiter(this Process process)
         {
             var tcs = new TaskCompletionSource<int>();
+            
             process.EnableRaisingEvents = true;
 
             process.Exited += (s, e) => tcs.TrySetResult(process.ExitCode);