+++ Runtime/Awaiters.cs
@@ -7,24 +7,11 @@ namespace GameLovers.AsyncAwait
 {
     public static class Awaiters
     {
-        private readonly static WaitForUpdate _waitForUpdate = new WaitForUpdate();
-        private readonly static WaitForFixedUpdate _waitForFixedUpdate = new WaitForFixedUpdate();
-        private readonly static WaitForEndOfFrame _waitForEndOfFrame = new WaitForEndOfFrame();
+        public static WaitForUpdate NextFrame { get; } = new WaitForUpdate();
 
-        public static WaitForUpdate NextFrame
-        {
-            get { return _waitForUpdate; }
-        }
+        public static WaitForFixedUpdate FixedUpdate { get; } = new WaitForFixedUpdate();
 
-        public static WaitForFixedUpdate FixedUpdate
-        {
-            get { return _waitForFixedUpdate; }
-        }
-
-        public static WaitForEndOfFrame EndOfFrame
-        {
-            get { return _waitForEndOfFrame; }
-        }
+        public static WaitForEndOfFrame EndOfFrame { get; } = new WaitForEndOfFrame();
 
         public static WaitForSeconds Seconds(float seconds)
         {