+++ Runtime/IEnumeratorAwaitExtensions.cs
@@ -66,12 +66,6 @@ namespace GameLovers.AsyncAwait
             return awaiter;
         }
 
-        // Return itself so you can do things like (await new WWW(url)).bytes
-        public static SimpleCoroutineAwaiter<WWW> GetAwaiter(this WWW instruction)
-        {
-            return GetAwaiterReturnSelf(instruction);
-        }
-
         public static SimpleCoroutineAwaiter<AssetBundle> GetAwaiter(this AssetBundleCreateRequest instruction)
         {
             var awaiter = new SimpleCoroutineAwaiter<AssetBundle>();