+++ Runtime/UnityWebRequestAwaiter.cs
@@ -1,6 +1,3 @@
-using System;
-using System.Runtime.CompilerServices;
-using UnityEngine;
 using UnityEngine.Networking;
 
 // ReSharper disable once CheckNamespace
@@ -10,41 +7,12 @@ namespace GameLovers.AsyncAwait
 	/// <summary>
 	/// Allows to use <seealso cref="UnityWebRequest"/> with async/await mechanism
 	/// </summary>
-	public class UnityWebRequestAwaiter : INotifyCompletion
+	public class UnityWebRequestAwaiter : AsyncOperationAwaiter
 	{
-		private readonly UnityWebRequestAsyncOperation _asyncOp;
-		private Action _continuation;
-		private bool _invoked;
+		public new UnityWebRequestAsyncOperation AsyncOperation => (UnityWebRequestAsyncOperation) _asyncOp;
 
-		public UnityWebRequestAsyncOperation AsyncOperation => _asyncOp;
-		
-		private UnityWebRequestAwaiter() {}
-
-		public UnityWebRequestAwaiter(UnityWebRequestAsyncOperation asyncOp)
+		public UnityWebRequestAwaiter(UnityWebRequestAsyncOperation asyncOp) : base(asyncOp)
 		{
-			_asyncOp = asyncOp;
-			asyncOp.completed += OnRequestCompleted;
-		}
-
-		public bool IsCompleted => _asyncOp.isDone;
-
-		public void GetResult() { }
-
-		public void OnCompleted(Action continuation)
-		{
-			_continuation = continuation;
-
-			if (_invoked)
-			{
-				_continuation();
-			}
-		}
-
-		private void OnRequestCompleted(AsyncOperation obj)
-		{
-			_invoked = true;
-			
-			_continuation?.Invoke();
 		}
 
 		public static implicit operator UnityWebRequestAwaiter(UnityWebRequestAsyncOperation asyncOp)