+++ CHANGELOG.md
@@ -4,10 +4,15 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.0] - 2020-07-26
+
+- Verified Package
+- Added *AsyncOperationAwaiter*. Now it is possible to do async await calls to Unity's *AsyncOperations*
+
 ## [0.2.0] - 2020-04-06
 
-- Added UnityWebRequestAwaiter. Now it is possible to do async await calls to UnityWebRequests
-- Added sample example to UnityWebRequestAwaiter
+- Added *UnityWebRequestAwaiter*. Now it is possible to do async await calls to *UnityWebRequests*
+- Added sample example to *UnityWebRequestAwaiter*
 
 ## [0.1.0] - 2020-01-06